MB-60625: Return size_t after accumulating to prevent underflow
Previously, the size of an arena is accunmulated into a 32 bit integer.
This meant that an underflow would occur after 4GB of memory had been
used. Instead, we correctly accumulate into a 64 bit integer and prevent
an underflow from occuring.

Reviewed-on: https://review.couchbase.org/c/platform/+/204864
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>

+++ src/je_arena_simple_tracker.cc
@@ -68,7 +68,7 @@ void JEArenaSimpleTracker::clientUnregistered(const ArenaMallocClient& client) {
 size_t JEArenaSimpleTracker::getPreciseAllocated(
         const ArenaMallocClient& client) {
     const auto& clientData = allocated.at(client.index);
-    return std::accumulate(clientData.begin(), clientData.end(), 0);
+    return std::accumulate(clientData.begin(), clientData.end(), size_t(0));
 }
 
 size_t JEArenaSimpleTracker::getPreciseAllocated(
